feat: Add EventDeliveryStatistics and Helpers classes for event delivery metrics and safe HTTP reading
refactor: Remove redundant BenchState and CountCache classes from Program.cs

feat: Introduce SerializableMultiProjectionStateEnvelope and SerializableMultiProjectionStateOffloaded for snapshot handling

fix: Update package versions to 1.0.2-preview10 across multiple projects

test: Implement AzuriteCollection and AzuriteTestFixture for Azure Blob Storage unit tests

test: Update IGuidTagGroupTests and TagPatternMatchingTests to use interface methods for tag content retrieval
