+++ internalUsages/Dcb.Domain/Queries/GetWeatherForecastCountQuery.cs
@@ -31,13 +31,3 @@ public record GetWeatherForecastCountQuery :
         ));
     }
 }
-
-/// <summary>
-/// Result containing weather forecast counts
-/// </summary>
-[GenerateSerializer]
-public record WeatherForecastCountResult(
-    [property: Id(0)] int SafeVersion,
-    [property: Id(1)] int UnsafeVersion,
-    [property: Id(2)] int TotalCount
-);
