+++ internalUsages/Dcb.Domain/Queries/WeatherForecastCountResult.cs
@@ -1,0 +1,12 @@
+using Orleans;
+namespace Dcb.Domain.Queries;
+
+/// <summary>
+/// Result containing weather forecast counts
+/// </summary>
+[GenerateSerializer]
+public record WeatherForecastCountResult(
+    [property: Id(0)] int SafeVersion,
+    [property: Id(1)] int UnsafeVersion,
+    [property: Id(2)] int TotalCount
+);
