+++ internalUsages/Dcb.Domain/Weather/WeatherForecastTag.cs
@@ -6,8 +6,6 @@ public record WeatherForecastTag(Guid ForecastId) : IGuidTagGroup<WeatherForecas
     public static string TagGroupName => "WeatherForecast";
 
     public bool IsConsistencyTag() => true;
-    public string GetTagContent() => ForecastId.ToString();
-
     public static WeatherForecastTag FromContent(string content)
     {
         if (Guid.TryParse(content, out var forecastId))
@@ -16,9 +14,5 @@ public record WeatherForecastTag(Guid ForecastId) : IGuidTagGroup<WeatherForecas
         }
         throw new ArgumentException($"Invalid forecast ID format: {content}");
     }
-
-    /// <summary>
-    ///     Get the GUID identifier for this tag
-    /// </summary>
     public Guid GetId() => ForecastId;
 }
