+++ internalUsages/DcbOrleans.Benchmark/CountCache.cs
@@ -1,0 +1,5 @@
+sealed class CountCache
+{
+    public int? Value { get; set; }
+    public DateTime? LastFetchedUtc { get; set; }
+}
