+++ internalUsages/DcbOrleans.Benchmark/Program.cs
@@ -614,82 +614,6 @@ static async Task RunAsync(string apiBase, BenchState state)
     }
 }
 
-sealed class BenchState
-{
-    public volatile bool IsRunning;
-    public int Total;
-    public int Concurrency;
-    public int Created;
-    public int Updated;
-    public int Errors;
-    public int QueryPageCount;
-    public bool StopOnError;
-    public bool UseSingle;
-    public string? Mode;
-    public bool HasRun;
-    public bool Canceled => _cts?.IsCancellationRequested == true;
-    public string? LastError;
-    private System.Diagnostics.Stopwatch _sw = new();
-    private CancellationTokenSource? _cts;
-    public TimeSpan Elapsed => _sw.Elapsed;
-    public CancellationToken Token => _cts?.Token ?? CancellationToken.None;
-    public void Reset(int total, int concurrency)
-    {
-        Total = total; Concurrency = concurrency; Created = Updated = Errors = QueryPageCount = 0;
-        StopOnError = false; LastError = null; _sw.Reset(); _cts?.Dispose(); _cts = new();
-    }
-    public void Start() => _sw.Start();
-    public void Stop() => _sw.Stop();
-    public void Cancel() => _cts?.Cancel();
-
-    private readonly Dictionary<string, CountCache> _countCache = new(StringComparer.OrdinalIgnoreCase);
-    public CountCache GetOrCreateCountCache(string mode)
-    {
-        if (!_countCache.TryGetValue(mode, out var c))
-        {
-            c = new CountCache();
-            _countCache[mode] = c;
-        }
-        return c;
-    }
-}
-
-sealed class CountCache
-{
-    public int? Value { get; set; }
-    public DateTime? LastFetchedUtc { get; set; }
-}
-
 // Helper class for deserializing weather count response
-public class WeatherCountResponse
-{
-  public int safeVersion { get; set; }
-  public int unsafeVersion { get; set; }
-  public int totalCount { get; set; }
-}
 
 // Helper class for deserializing event delivery statistics
-public class EventDeliveryStatistics
-{
-    public int totalUniqueEvents { get; set; }
-    public long totalDeliveries { get; set; }
-    public long duplicateDeliveries { get; set; }
-    public int eventsWithMultipleDeliveries { get; set; }
-    public int maxDeliveryCount { get; set; }
-    public double averageDeliveryCount { get; set; }
-    public int? streamUniqueEvents { get; set; }
-    public long? streamDeliveries { get; set; }
-    public int? catchUpUniqueEvents { get; set; }
-    public long? catchUpDeliveries { get; set; }
-}
-
-static class Helpers
-{
-    public static async Task<string> SafeReadAsync(HttpResponseMessage res, CancellationToken ct)
-    {
-        try { return await res.Content.ReadAsStringAsync(ct); }
-        catch { return string.Empty; }
-    }
-    public static string Trim(string s)
-        => string.IsNullOrEmpty(s) ? s : (s.Length > 256 ? s.Substring(0, 256) + "..." : s);
-}
