+++ internalUsages/DcbOrleans.Benchmark/WeatherCountResponse.cs
@@ -1,0 +1,6 @@
+public class WeatherCountResponse
+{
+    public int safeVersion { get; set; }
+    public int unsafeVersion { get; set; }
+    public int totalCount { get; set; }
+}
