+++ src/Sekiban.Dcb/MultiProjections/SafeProjection.cs
@@ -1,0 +1,3 @@
+namespace Sekiban.Dcb.MultiProjections;
+
+public readonly record struct SafeProjection<T>(T State, string SafeLastSortableUniqueId, int Version) where T : IMultiProjectionPayload;
