+++ src/Sekiban.Dcb/Snapshots/SerializableMultiProjectionStateOffloaded.cs
@@ -15,14 +15,3 @@ public sealed record SerializableMultiProjectionStateOffloaded(
     bool IsCatchedUp,
     bool IsSafeState,
     long PayloadLength);
-
-/// <summary>
-///     Envelope that contains either inline snapshot or offloaded snapshot reference.
-/// </summary>
-public sealed record SerializableMultiProjectionStateEnvelope(
-    bool IsOffloaded,
-    Sekiban.Dcb.MultiProjections.SerializableMultiProjectionState? InlineState,
-    SerializableMultiProjectionStateOffloaded? OffloadedState)
-{
-}
-
