+++ src/Sekiban.Dcb/Tags/IGuidTagGroup.cs
@@ -7,24 +7,10 @@ namespace Sekiban.Dcb.Tags;
 /// <typeparam name="TTagGroup">The concrete tag type</typeparam>
 public interface IGuidTagGroup<TTagGroup> : ITagGroup<TTagGroup> where TTagGroup : IGuidTagGroup<TTagGroup>
 {
-
-    // Re-declare static abstract members to avoid CS8920
-    static abstract new string TagGroupName { get; }
+    string ITag.GetTagContent() => GetId().ToString("D");
     /// <summary>
     ///     Get the GUID identifier for this tag
     /// </summary>
     /// <returns>The GUID identifier</returns>
     Guid GetId();
-}
-
-public interface IStringTagGroup<TTagGroup> : ITagGroup<TTagGroup> where TTagGroup : IStringTagGroup<TTagGroup>
-{
-
-    // Re-declare static abstract members to avoid CS8920
-    static abstract new string TagGroupName { get; }
-    /// <summary>
-    ///     Get the string identifier for this tag
-    /// </summary>
-    /// <returns>The string identifier</returns>
-    string GetId();
 }
