+++ src/Sekiban.Pure.Postgres/Sekiban.Pure.Postgres.csproj
@@ -5,12 +5,12 @@
         <Nullable>enable</Nullable>
         <LangVersion>preview</LangVersion>
         <PackageId>Sekiban.Pure.Postgres</PackageId>
-        <Version>1.0.2-preview09</Version>
+        <Version>1.0.2-preview10</Version>
         <Authors>J-Tech Group</Authors>
         <Company>J-Tech-Japan</Company>
         <PackageDescription>Sekiban - Pure Event Sourcing Framework PostgreSQL Integration</PackageDescription>
         <RepositoryUrl>https://github.com/J-Tech-Japan/Sekiban</RepositoryUrl>
-        <PackageVersion>1.0.2-preview09</PackageVersion>
+        <PackageVersion>1.0.2-preview10</PackageVersion>
         <Description>Multi Blob saving and reading issue resolved</Description>
         <AssemblyName>Sekiban.Pure.Postgres</AssemblyName>
         <RootNamespace>Sekiban.Pure.Postgres</RootNamespace>
@@ -33,7 +33,7 @@
         <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="9.0.8"/>
         <PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="9.0.8"/>
         <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="9.0.4"/>
-        <PackageReference Include="Sekiban.Pure.AspNetCore" Version="1.0.2-preview09"/>
+        <PackageReference Include="Sekiban.Pure.AspNetCore" Version="1.0.2-preview10"/>
         <PackageReference Include="System.Configuration.ConfigurationManager" Version="9.0.8"/>
         <PackageReference Include="Microsoft.Sbom.Targets" Version="4.1.2">
             <PrivateAssets>all</PrivateAssets>
