+++ tests/Sekiban.Dcb.BlobStorage.AzureStorage.Unit/AzuriteCollection.cs
@@ -1,0 +1,7 @@
+using Xunit;
+namespace Sekiban.Dcb.BlobStorage.AzureStorage.Unit;
+
+[CollectionDefinition("AzuriteCollection")]
+public class AzuriteCollection : ICollectionFixture<AzuriteTestFixture>
+{
+}
